+++ benches/library_benchmark.rs
@@ -1,6 +1,6 @@
-use engine::engine::move_generation::perft::init_test_func;
-use engine::engine::move_generation::perft::Stats;
-use engine::engine::shared::helper_func::const_utility::FEN_START;
+use engine::engine::board::perft::init_test_func;
+use engine::engine::board::perft::Stats;
+use engine::engine::misc::const_utility::FEN_START;
 use iai_callgrind::library_benchmark;
 use iai_callgrind::library_benchmark_group;
 use iai_callgrind::main;
