+++ src/engine/board/structures/mod.rs
@@ -1,0 +1,9 @@
+
+pub mod castling;
+pub mod square;
+pub mod state;
+pub mod moves;
+pub mod board;
+pub mod color;
+pub mod piece;
+
