+++ src/engine/misc/bit_pos_utility.rs
