+++ src/engine/misc/const_utility.rs
