+++ src/engine/misc/mod.rs
@@ -1,6 +1,7 @@
 pub mod bit_pos_utility;
 pub mod bitboard;
 pub mod const_utility;
+pub mod directions;
 pub mod generate_key_utility;
 pub mod play_chess_utility;
 pub mod print_utility;
