+++ src/engine/mod.rs
@@ -1,7 +1,6 @@
+pub mod board;
 pub mod evaluation;
-pub mod move_generation;
+pub mod misc;
 pub mod move_generator;
 pub mod protocols;
 pub mod search;
-pub mod shared;
-pub mod board;
