--- src/engine/move_generation/mod.rs
@@ -1,3 +1,0 @@
-pub mod make_move;
-pub mod mv_gen;
-pub mod perft;
