+++ src/engine/move_generator/bishop.rs
@@ -1,6 +1,7 @@
 use super::generated::bishop::{BISHOP_BASE, BISHOP_LOOKUP, BISHOP_MASKS};
 use super::utility::pext;
 
+// TODO: ADD CONSTANTS TO A BETTER PLACE
 const WHITE_SQUARES: u64 = 0b1010101010101010101010101010101010101010101010101010101010101010;
 const BLACK_SQUARES: u64 = 0b0101010101010101010101010101010101010101010101010101010101010101;
