+++ src/engine/move_generator/rays.rs
@@ -1,5 +1,5 @@
-use crate::engine::shared::helper_func::bitboard::BitboardTrait;
-use crate::engine::shared::structures::directions::*;
+use crate::engine::misc::bitboard::BitboardTrait;
+use crate::engine::misc::directions::Dir;
 
 pub fn first_hit(dir: Dir, bitboard: u64) -> Option<usize> {
     if bitboard == 0 {
@@ -27,11 +27,10 @@ pub fn blocked_ray_att(dir: Dir, ray_family: &[u64; 64], ray: u64, own: u64, ene
 
 #[cfg(test)]
 mod tests {
+    use crate::engine::misc::bit_pos_utility::*;
+    use crate::engine::misc::bitboard::BitboardTrait;
+    use crate::engine::misc::print_utility::print_bitboard;
     use crate::engine::move_generator::generated::rays::RAYS_LOOKUP;
-    use crate::engine::shared::helper_func::bit_pos_utility::*;
-    use crate::engine::shared::helper_func::bitboard::BitboardTrait;
-    use crate::engine::shared::helper_func::print_utility::print_bitboard;
-    use crate::engine::shared::structures::directions::Dir;
 
     use super::*;
