+++ src/engine/search/mod.rs
@@ -1,3 +1,2 @@
 pub mod searcher;
-pub mod time;
 pub mod transposition_table;
