--- src/engine/shared/mod.rs
@@ -1,2 +1,0 @@
-pub mod helper_func;
-pub mod structures;
