--- src/engine/shared/structures/mod.rs
@@ -1,6 +1,0 @@
-pub mod castling_struct;
-pub mod color;
-pub mod directions;
-pub mod internal_move;
-pub mod piece;
-pub mod square;
